3 CODEC adapter tests are passed.
Also fixed typo ( Adaptor -> Adapter )

+++ drivers/codec/tlv320aic3204adapterstrategy.hpp
@@ -26,9 +26,7 @@ class Tlv320aic3204AdapterStrategy {
    */
   Tlv320aic3204AdapterStrategy(
       murasaki::I2cMasterStrategy *controller,  // I2C master controller
-      unsigned int device_addr
-
-  );
+      u_int8_t device_addr);
 
   /**
    * \brief Set the CODEC register page to page_number
